This was what I thought I checked in in revision 717 - stoopid textmate letting me clicking on one file and then going ahead and committing another.

+++ trunk/www/blog.seagul.co.uk/bin/export_articles_to_filesystem.rb
@@ -19,11 +19,11 @@ Article.find(:all).each do |article|
   filename = "#{published_at_for_filename}-#{article.permalink}.yml"
   filepath = File.join(File.dirname(__FILE__), '..', 'articles', filename)
 
-  # if File.exists?(filepath)
-    # warn "WARNING: Cannot overwrite existing article called: #{filename}"
-  # else
+  if File.exists?(filepath)
+    warn "WARNING: Cannot overwrite existing article called: #{filename}"
+  else
     File.open(filepath, 'w') { |f| f.puts(article_attributes.to_yaml) }
     puts "Created article: #{filename}"
-  # end
+  end
   
 end
